+++ 02week/ticTacToe.js
@@ -1,0 +1,143 @@
+'use strict';
+
+const assert = require('assert');
+const readline = require('readline');
+const rl = readline.createInterface({
+  input: process.stdin,
+  output: process.stdout
+});
+let board = [
+  [' ', ' ', ' '],
+  [' ', ' ', ' '],
+  [' ', ' ', ' ']
+];
+
+let playerTurn = 'X';
+
+function printBoard() {
+  console.log('   0  1  2');
+  console.log('0 ' + board[0].join(' | '));
+  console.log('  ---------');
+  console.log('1 ' + board[1].join(' | '));
+  console.log('  ---------');
+  console.log('2 ' + board[2].join(' | '));
+}
+
+function horizontalWin() {
+  // Your code here
+  if (board === [['X', 'X', 'X'], [' ', ' ', ' '], [' ', ' ', ' ']]
+    || board === [[' ', ' ', ' '], ['X', 'X', 'X'], [' ', ' ', ' ']]
+    || board === [[' ', ' ', ' '], [' ', ' ', ' '], ['X', 'X', 'X']]
+    || board === [['0', '0', '0'], [' ', ' ', ' '], [' ', ' ', ' ']]
+    || board === [[' ', ' ', ' '], ['0', '0', '0'], [' ', ' ', ' ']]
+    || board === [[' ', ' ', ' '], [' ', ' ', ' '], ['0', '0', '0']]
+  ) {
+    return true;
+  }
+}
+
+function verticalWin() {
+  // Your code here
+  if (board == [['X', ' ', ' '], ['X', ' ', ' '], ['X', ' ', ' ']]
+    || board == [[' ', 'X', ' '], [' ', 'X', ' '], [' ', 'X', ' ']]
+    || board == [[' ', ' ', 'X'], [' ', ' ', 'X'], [' ', ' ', 'X']]
+    || board == [['0', ' ', ' '], ['0', ' ', ' '], ['0', ' ', ' ']]
+    || board == [[' ', '0', ' '], [' ', '0', ' '], [' ', '0', ' ']]
+    || board == [[' ', ' ', '0'], [' ', ' ', '0'], [' ', ' ', '0']]
+  ) {
+    return true;
+  }
+}
+
+function diagonalWin() {
+  // Your code here
+  if (board == [['X', ' ', ' '], [' ', 'X', ' '], [' ', ' ', 'X']]
+    || board == [[' ', ' ', 'X'], [' ', 'X', ' '], ['X', ' ', ' ']]
+    || board == [['0', ' ', ' '], [' ', '0', ' '], [' ', ' ', '0']]
+    || board[[' ', ' ', '0'], [' ', '0', ' '], ['0', ' ', ' ']]
+  ) {
+    return true;
+  }
+}
+
+function checkForWin() {
+  // Your code here
+  if (horizontalWin() || verticalWin() || diagonalWin()) {
+    return true;
+  }
+}
+
+function ticTacToe(row, column) {
+  // Your code here
+  if (playerTurn === 'X') {
+    board[row][column] = playerTurn;
+    if (checkForWin()) {
+      return "winner is: " + playerTurn;
+    } else {
+      playerTurn = '0';
+    }
+  }
+  else {
+    board[row][column] = playerTurn;
+    if (checkForWin()) {
+      return "winner is: " + playerTurn;
+    } else {
+      playerTurn = 'X';
+    }
+  }
+}
+
+function getPrompt() {
+  printBoard();
+  console.log("It's Player " + playerTurn + "'s turn.");
+  rl.question('row: ', (row) => {
+    rl.question('column: ', (column) => {
+      try {
+        ticTacToe(row, column);
+      }
+      catch(err) {
+        console.log("WHAT DID YOU DO!!!!!! PLEASE ENTER CORRECT VALUE")
+      }
+      
+      getPrompt();
+    });
+  });
+
+}
+
+
+
+// Tests
+
+if (typeof describe === 'function') {
+
+  describe('#ticTacToe()', () => {
+    it('should place mark on the board', () => {
+      ticTacToe(1, 1);
+      assert.deepEqual(board, [[' ', ' ', ' '], [' ', 'X', ' '], [' ', ' ', ' ']]);
+    });
+    it('should alternate between players', () => {
+      ticTacToe(0, 0);
+      assert.deepEqual(board, [['O', ' ', ' '], [' ', 'X', ' '], [' ', ' ', ' ']]);
+    });
+    it('should check for vertical wins', () => {
+      board = [[' ', 'X', ' '], [' ', 'X', ' '], [' ', 'X', ' ']];
+      assert.equal(verticalWin(), true);
+    });
+    it('should check for horizontal wins', () => {
+      board = [['X', 'X', 'X'], [' ', ' ', ' '], [' ', ' ', ' ']];
+      assert.equal(horizontalWin(), true);
+    });
+    it('should check for diagonal wins', () => {
+      board = [['X', ' ', ' '], [' ', 'X', ' '], [' ', ' ', 'X']];
+      assert.equal(diagonalWin(), true);
+    });
+    it('should detect a win', () => {
+      assert.equal(checkForWin(), true);
+    });
+  });
+} else {
+
+  getPrompt();
+
+}
